added ElasticsearchWriteOptions, fixed elasticsearch-common structure and tests
