+++ v2/bigquery-to-elasticsearch/src/test/java/com/google/cloud/teleport/v2/templates/BigQueryToElasticsearchTest.java
@@ -52,9 +52,9 @@ public class BigQueryToElasticsearchTest {
         PipelineOptionsFactory.create()
             .as(BigQueryToElasticsearch.BigQueryToElasticsearchReadOptions.class);
 
-    options.setNodeAddresses("http://my-node");
-    options.setIndex("test");
-    options.setDocumentType("_doc");
+    options.setTargetNodeAddresses("http://my-node");
+    options.setWriteIndex("test");
+    options.setWriteDocumentType("_doc");
     options.setInputTableSpec("my-project:my-dataset.my-table");
     options.setQuery(null);
 
@@ -88,9 +88,9 @@ public class BigQueryToElasticsearchTest {
         PipelineOptionsFactory.create()
             .as(BigQueryToElasticsearch.BigQueryToElasticsearchReadOptions.class);
 
-    options.setNodeAddresses("http://my-node");
-    options.setIndex("test");
-    options.setDocumentType("_doc");
+    options.setTargetNodeAddresses("http://my-node");
+    options.setWriteIndex("test");
+    options.setWriteDocumentType("_doc");
     options.setInputTableSpec(null);
     options.setQuery(null);
